build controller: set privileged=false
Set the "privileged" flag on docker/sti/custom build pods to false, and
put CAP_MKNOD and CAP_KILL in the list of capabilities to drop.

Make the permissions we set for ConfigMap and Secret volumes parameters
that we can pass down to our helper functions which add their volumes to
the pod spec, and set them so that their contents are group-readable by
their owners.

Mount an emptyDir volume at /var/run/containers in
setupContainersStorage(), so that we can ensure its permissions will be
set so that the build user will always be able to write to it.

Move the emptyDir volume from /var/lib/containers/storage to
/var/lib/containers, so that the build user can create the storage
directory inside of it with exactly the permissions that the build user
wants.

Add helper functions for running build pods as the first UID in the
namespace's UID range, using the first GID in the namespace's set of GID
ranges as the pod's fsGroup, and with /etc/subuid and /etc/subgid in our
containers populated using the UID and GID ranges from the namespace.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>